Add canonicalize_file_name from libc
The .net core Path class can't always do this correctly by itself. I'm using this function in a project, so I figured I'd contribute it here as well.

Confirmed working on x64 and arm64. I don't have am arm32 system to check it on, but the API is the same for this function, so it should work there as well.